backendpy/requirements.txt
Flask==2.0.2
Flask-Cors==3.0.10
Flask-PyMongo==2.3.0
python-dotenv==0.19.2
openai==0.10.2
Werkzeug==2.2.2